split contact picker cell into its own file and make it safe
